Share and LoadPlot commands working
Now you can share your plot by using /p share [name]
The name is optional, but encouraged. Without it,
a default, somewhat hard to remember name is given to it
so do yourself a favor, and give it a name. :)

/p loadplot <player name> <plot id>
this can be used with either the id of the plot share
or the name given to it by the player. Please use names!
You must also include the player's name as indicated
so we know whose plot share you're referring to.